Initial take at compat system. Skeleton code with partially converted `DataCount`.
