+++ src/compat/base/base-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/base-mixin';
