+++ src/compat/base/bubble-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/bubble-mixin';
