+++ src/compat/base/cap-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/cap-mixin';
