+++ src/compat/base/color-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/color-mixin';
