+++ src/compat/base/coordinate-grid-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/coordinate-grid-mixin';
