+++ src/compat/base/d3.box.ts
@@ -1,0 +1,1 @@
+export * from '../../base/d3.box';
