+++ src/compat/base/margin-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/margin-mixin';
