+++ src/compat/base/stack-mixin.ts
@@ -1,0 +1,1 @@
+export * from '../../base/stack-mixin';
