+++ src/compat/charts/bar-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/bar-chart';
