+++ src/compat/charts/box-plot.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/box-plot';
