+++ src/compat/charts/bubble-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/bubble-chart';
