+++ src/compat/charts/bubble-overlay.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/bubble-overlay';
