+++ src/compat/charts/cbox-menu.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/cbox-menu';
