+++ src/compat/charts/composite-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/composite-chart';
