+++ src/compat/charts/data-grid.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/data-grid';
