+++ src/compat/charts/data-table.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/data-table';
