+++ src/compat/charts/geo-choropleth-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/geo-choropleth-chart';
