+++ src/compat/charts/heat-map.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/heat-map';
