+++ src/compat/charts/html-legend.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/html-legend';
