+++ src/compat/charts/legend.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/legend';
