+++ src/compat/charts/line-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/line-chart';
