+++ src/compat/charts/number-display.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/number-display';
