+++ src/compat/charts/pie-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/pie-chart';
