+++ src/compat/charts/row-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/row-chart';
