+++ src/compat/charts/scatter-plot.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/scatter-plot';
