+++ src/compat/charts/select-menu.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/select-menu';
