+++ src/compat/charts/series-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/series-chart';
