+++ src/compat/charts/sunburst-chart.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/sunburst-chart';
