+++ src/compat/charts/text-filter-widget.ts
@@ -1,0 +1,1 @@
+export * from '../../charts/text-filter-widget';
