+++ src/compat/core/bad-argument-exception.ts
@@ -1,0 +1,1 @@
+export * from '../../core/bad-argument-exception';
