+++ src/compat/core/chart-registry.ts
@@ -1,0 +1,1 @@
+export * from '../../core/chart-registry';
