+++ src/compat/core/config.ts
@@ -1,0 +1,1 @@
+export * from '../../core/config';
