+++ src/compat/core/constants.ts
@@ -1,0 +1,1 @@
+export * from '../../core/constants';
