+++ src/compat/core/core.ts
@@ -1,0 +1,1 @@
+export * from '../../core/core';
