+++ src/compat/core/events.ts
@@ -1,0 +1,1 @@
+export * from '../../core/events';
