+++ src/compat/core/invalid-state-exception.ts
@@ -1,0 +1,1 @@
+export * from '../../core/invalid-state-exception';
