+++ src/compat/core/logger.ts
@@ -1,0 +1,1 @@
+export * from '../../core/logger';
