+++ src/compat/core/printers.ts
@@ -1,0 +1,1 @@
+export * from '../../core/printers';
