+++ src/compat/core/types.ts
@@ -1,0 +1,1 @@
+export * from '../../core/types';
