+++ src/compat/core/utils.ts
@@ -1,0 +1,1 @@
+export * from '../../core/utils';
