+++ src/compat/index-compat.ts
@@ -1,0 +1,46 @@
+export * from './core/bad-argument-exception';
+export * from './core/chart-registry';
+export * from './core/config';
+export * from './core/constants';
+export * from './core/core';
+export * from './core/events';
+export * from './core/filters';
+export * from './core/invalid-state-exception';
+export * from './core/logger';
+export * from './core/printers';
+export * from './core/units';
+export * from './core/utils';
+
+export * from './base/base-mixin';
+export * from './base/bubble-mixin';
+export * from './base/cap-mixin';
+export * from './base/color-mixin';
+export * from './base/coordinate-grid-mixin';
+export * from './base/d3.box';
+export * from './base/margin-mixin';
+export * from './base/stack-mixin';
+
+export * from './charts/bar-chart';
+export * from './charts/box-plot';
+export * from './charts/bubble-chart';
+export * from './charts/bubble-overlay';
+export * from './charts/cbox-menu';
+export * from './charts/composite-chart';
+export * from './charts/data-count';
+export * from './charts/data-grid';
+export * from './charts/data-table';
+export * from './charts/geo-choropleth-chart';
+export * from './charts/heat-map';
+export * from './charts/html-legend';
+export * from './charts/legend';
+export * from './charts/line-chart';
+export * from './charts/number-display';
+export * from './charts/pie-chart';
+export * from './charts/row-chart';
+export * from './charts/scatter-plot';
+export * from './charts/select-menu';
+export * from './charts/series-chart';
+export * from './charts/sunburst-chart';
+export * from './charts/text-filter-widget';
+
+export {version} from '../index-with-version';
